Adding support for multiple single HTTP header entries.
When parsing a HTTP response that has multiple HTTP headers as single
entries, as opposed to a single header with a comma separated list of
values, only the last value was added to the dictionary of headers.

This addition checks if the header name has already been added to the
dictionary, and if so, appends the value to the already existing key's
value.