Updating handling of '*' access-control-allow-origin
Motivation:
If a CORS enabled server returns the wildcard origin '*' then the
access-control-allow-credentials should not be set. If the server does
provied a actual origin in the access-control-allow-origin header then
access-control-allow-credentials must be true.

Currently there are some inconsitencies in the test in this regard.

Modifications:
Updated the test with checks to see is check for a '*' origin and verify
that access-control-allow-credentials is not set. Also only checking
access-control-allow-credentials when there is an explicit origin.

Result:
All test should still pass.